Adding many game related component and also added the player controller.

+++ mario/src/main/java/components/FontRender.java
@@ -9,7 +9,7 @@ public class FontRender extends Component {
     }
   }
   @Override
-  public void update(double dt) {
+  public void update(float dt) {
 
   }
 }
