+++ mario/src/main/java/components/TranslateGizmo.java
@@ -11,10 +11,12 @@ public class TranslateGizmo extends Gizmo {
 
 
   @Override
-  public void editorUpdate(double dt) {
+  public void editorUpdate(float dt) {
 
     if (activeGameObject != null) {
+      //System.out.println("there is an active go!");
       if (xAxisActive || yAxisActive) {
+        //System.out.println("there is an active gizmo!");
         activeGameObject.getTransform().setPositionY(activeGameObject.getTransform().getPosition().y - MouseListener.getWorldDeltaY());
         activeGameObject.getTransform().setPositionX(activeGameObject.getTransform().getPosition().x - MouseListener.getWorldDeltaX());
       }
